+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## Unreleashed
+## [Unreleased]
 
 - [TD-1747] Structures upload is not creating relation between structures and fields when including version number
+- [TD-1758] Structures with 'field' class are no indexed on create/upload
 
 ## [2.19.0] 2019-05-14